bfsdfs, triad changed to 64bit

+++ snap-core/flow.h
@@ -4,8 +4,8 @@ namespace TSnap {
 const TStr CapAttrName = "capacity";
 
 /// Returns the maximum integer valued flow in the network \c Net from source \c SrcNId to sink \c SnkNId. ##TSnap::GetMaxFlowIntEK
-int GetMaxFlowIntEK (PNEANet &Net, const int &SrcNId, const int &SnkNId);
+int64 GetMaxFlowIntEK (PNEANet &Net, const int64 &SrcNId, const int64 &SnkNId);
 /// Returns the maximum integer valued flow in the network \c Net from source \c SrcNId to sink \c SnkNId. ##TSnap::GetMaxFlowIntEK
-int GetMaxFlowIntPR (PNEANet &Net, const int &SrcNId, const int &SnkNId);
+int64 GetMaxFlowIntPR (PNEANet &Net, const int64 &SrcNId, const int64 &SnkNId);
 
 };
